add quota routes
